Match dataset based on both topic and record type (#175)
When consuming messages from topics with multiple schemas, we would like to be able to choose dataset based on both - topic name and record type.
Because patterns have to be unique and new record types might appear in the topic, I introduced a new setting that allows users to specify which dataset should be used when none of the `dataset` patterns matches. When this parameter is not set or set to null, connector will behave just like before and throw an exception.